colors are in a separate file now cutting plane work

+++ pyNastran/gui/qt_files/gui_attributes.py
@@ -767,7 +767,7 @@ class GuiAttributes(object):
 
         Parameters
         ----------
-        coord_id : float
+        coord_id : int
             the coordinate system id
         dim_max : float
             the max model dimension; 10% of the max will be used for the coord length
